Struct lyon_tessellation::geometry_builder::BuffersBuilder

pub struct BuffersBuilder<'l, VertexType: 'l, Input, Ctor: VertexConstructor<Input, VertexType>> { /* fields omitted */ }

A temporary view on a VertexBuffers object which facilitate the population of vertex and index data.

BuffersBuilders record the vertex offset from when they are created so that algorithms using them don't need to worry about offsetting indices if some geometry was added beforehand. This means that from the point of view of a BuffersBuilder user, the first added vertex is at always offset at the offset 0 and VertexBuilfer takes care of translating indices adequately.

Often, algorithms are built to generate vertex positions without knowledge of eventual other vertex attributes. The VertexConstructor does the translation from generic Input to VertexType. If your logic generates the actual vertex type directly, you can use the SimpleBuffersBuilder convenience typedef.

Methods

impl<'l, VertexType: 'l, Input, Ctor: VertexConstructor<Input, VertexType>> BuffersBuilder<'l, VertexType, Input, Ctor>

fn new(buffers: &'l mut VertexBuffers<VertexType>,
ctor: Ctor)
-> BuffersBuilder<'l, VertexType, Input, Ctor>