Struct lyon_tessellation::path_fill::FillTessellator

pub struct FillTessellator {
    pub _handle_intersections: bool,
    // some fields omitted
}

A Context object that can tessellate fill operations for complex paths.

The Tessellator API is not stable yet. For example it is not clear whether we will use separate Tessellator structs for some of the different configurations (vertex-aa, etc), or if evertything can be implemented with the same algorithm.

Fields
_handle_intersections: bool
Methods

impl FillTessellator

fn new() -> FillTessellator

Constructor.

fn tessellate_path<Iter, Output>(&mut self,
it: Iter,
options: &FillOptions,
output: &mut Output)
-> FillResult where Iter: Iterator<Item=FlattenedEvent>, Output: BezierGeometryBuilder<Vertex>

Compute the tessellation from a path iterator.

fn tessellate_events<Output>(&mut self,
events: &FillEvents,
options: &FillOptions,
output: &mut Output)
-> FillResult where Output: BezierGeometryBuilder<Vertex>

Compute the tessellation from pre-sorted events.

fn enable_logging(&mut self)

Enable some verbose logging during the tessellation, for debugging purposes.