lyon_tessellation
Struct EventQueue

pub struct EventQueue { /* private fields */ }

A queue of sorted events for the fill tessellator’s sweep-line algorithm.

Implementations

impl EventQueue

pub fn new() -> Self

pub fn with_capacity(cap: usize) -> Self

pub fn reset(&mut self)

pub fn from_path(tolerance: f32, path: impl Iterator<Item = PathEvent>) -> Self

Creates an EventQueue from an iterator of path event and a tolerance threshold.

The tolerance threshold is used for curve flattening approximation. See the Flattening and tolerance section of the crate documentation.

pub fn from_path_with_ids( tolerance: f32, sweep_orientation: Orientation, path: impl Iterator<Item = IdEvent>, positions: &impl PositionStore, ) -> Self

Creates an EventQueue from an an iterator over endpoint and control point ids, storage for the positions and, optionally, storage for custom endpoint attributes.

The tolerance threshold is used for curve flattening approximation. See the Flattening and tolerance section of the crate documentation.

pub fn into_builder(self) -> EventQueueBuilder

pub fn reserve(&mut self, n: usize)