lyon_tessellation
Struct StrokeOptions

#[non_exhaustive]
pub struct StrokeOptions {
    pub start_cap: LineCap,
    pub end_cap: LineCap,
    pub line_join: LineJoin,
    pub line_width: f32,
    pub miter_limit: f32,
    pub tolerance: f32,
}

Parameters for the tessellator.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

start_cap: LineCap

What cap to use at the start of each sub-path.

Default value: LineCap::Butt.

end_cap: LineCap

What cap to use at the end of each sub-path.

Default value: LineCap::Butt.

line_join: LineJoin

See the SVG specification.

Default value: LineJoin::Miter.

line_width: f32

Line width

Default value: StrokeOptions::DEFAULT_LINE_WIDTH.

miter_limit: f32

See the SVG specification.

Must be greater than or equal to 1.0. Default value: StrokeOptions::DEFAULT_MITER_LIMIT.

tolerance: f32

Maximum allowed distance to the path when building an approximation.

See Flattening and tolerance. Default value: StrokeOptions::DEFAULT_TOLERANCE.

Implementations

impl StrokeOptions

pub const MINIMUM_MITER_LIMIT: f32 = 1f32

Minimum miter limit as defined by the SVG specification.

See StrokeMiterLimitProperty

pub const DEFAULT_MITER_LIMIT: f32 = 4f32

Default miter limit as defined by the SVG specification.

See StrokeMiterLimitProperty

pub const DEFAULT_LINE_CAP: LineCap = LineCap::Butt

pub const DEFAULT_LINE_JOIN: LineJoin = LineJoin::Miter

pub const DEFAULT_LINE_WIDTH: f32 = 1f32

pub const DEFAULT_TOLERANCE: f32 = 0.100000001f32

pub const DEFAULT: Self = _

pub fn tolerance(tolerance: f32) -> Self

pub fn with_tolerance(self, tolerance: f32) -> Self

pub fn with_line_cap(self, cap: LineCap) -> Self

pub fn with_start_cap(self, cap: LineCap) -> Self

pub fn with_end_cap(self, cap: LineCap) -> Self

pub fn with_line_join(self, join: LineJoin) -> Self

pub fn with_line_width(self, width: f32) -> Self

pub fn with_miter_limit(self, limit: f32) -> Self