Struct lyon_tessellation::StrokeVertex

pub struct StrokeVertex<'a, 'b>(_, _);

Extra vertex information from the StrokeTessellator accessible when building vertices.

Implementations

impl<'a, 'b> StrokeVertex<'a, 'b>

pub fn position(&self) -> Point

The vertex position.

pub fn normal(&self) -> Vector

Normal at this vertex.

The length of the provided normal is such that displacing the vertex along it inflates the stroke by 2.0 (1.0 on each side).

pub fn position_on_path(&self) -> Point

Position of this vertex on the path, unaffected by the line width.

pub fn line_width(&self) -> f32

The line width at this vertex.

If a line width modifier is set via StrokeOptions::variable_line_width, the returned line width is equal to StrokeOptions::line_width multiplied by the line width modifier at this vertex.

pub fn advancement(&self) -> f32

How far along the path this vertex is.

pub fn side(&self) -> Side

Whether the vertex is on the positive or negative side of the path.

pub fn source(&self) -> VertexSource

Returns the source of this vertex.

pub fn interpolated_attributes(&mut self) -> Attributes<'_>

Computes and returns the custom attributes for this vertex.

The attributes are interpolated along the edges on which this vertex is. This can include multiple edges if the vertex is at an intersection.