lyon_tessellation::geometry_builder
Trait FillGeometryBuilder

pub trait FillGeometryBuilder: GeometryBuilder {
    // Required method
    fn add_fill_vertex(
        &mut self,
        vertex: FillVertex<'_>,
    ) -> Result<VertexId, GeometryBuilderError>;
}

A Geometry builder to interface with the FillTessellator.

Types implementing this trait must also implement the GeometryBuilder trait.

Required Methods

fn add_fill_vertex( &mut self, vertex: FillVertex<'_>, ) -> Result<VertexId, GeometryBuilderError>

Inserts a vertex, providing its position, and optionally a normal. Returns a vertex id that is only valid between begin_geometry and end_geometry.

This method can only be called between begin_geometry and end_geometry.